requirements.txt
# asyncio
fastapi>=0.68.0
# pydantic>=1.8.0
# python-dotenv>=0.20.0
# requests>=2.25.0
uvicorn>=0.15.0
psycopg2-binary